đt

+++ BeverageStoreManagement/BeverageStoreManagement/Models/Material.cs
@@ -12,7 +12,7 @@ namespace BeverageStoreManagement.Models
         private string nameMaterial;
         private string type;
         private string countUnit;
-        private int quantity;
+        private double quantity;
         private double purchasePrice;
         private byte[] image;
         private bool status;
@@ -22,14 +22,14 @@ namespace BeverageStoreManagement.Models
         public string NameMaterial { get => nameMaterial; set => nameMaterial = value; }
         public string Type { get => type; set => type = value; }
         public string CountUnit { get => countUnit; set => countUnit = value; }
-        public int Quantity { get => quantity; set => quantity = value; }
+        public double Quantity { get => quantity; set => quantity = value; }
         public double PurchasePrice { get => purchasePrice; set => purchasePrice = value; }
         public byte[] Image { get => image; set => image = value; }
         public bool Status { get => status; set => status = value; }
         public bool IsDelete { get => isDelete; set => isDelete = value; }
 
         public Material() { }
-        public Material(int idMaterial, string nameMaterial, string type, string countUnit, int quantity, double purchasePrice, byte[] image, bool status, bool isDelete)
+        public Material(int idMaterial, string nameMaterial, string type, string countUnit, double quantity, double purchasePrice, byte[] image, bool status, bool isDelete)
         {
             this.idMaterial = idMaterial;
             this.nameMaterial = nameMaterial;
